pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "screen-pilot"
version = "0.1.0"
description = "Give AI agents eyes and hands on your Linux desktop"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [{ name = "clintm" }]
keywords = ["mcp", "ai", "agent", "desktop", "automation", "screen", "vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.12",
    "Topic :: Desktop Environment",
]
dependencies = [
    "fastmcp>=3.0",
    "fastapi>=0.110",
    "uvicorn>=0.29",
    "typer>=0.12",
    "textual>=0.80",
    "python-ydotool>=1.0",
    "pixelmatch>=0.3",
    "Pillow>=10.0",
    "requests>=2.31",
    "tomli>=2.0; python_version < '3.11'",
    "tomli-w>=1.0",
]

[project.optional-dependencies]
vision = [
    "ultralytics>=8.4.0",
    "torch>=2.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
screen-pilot = "screen_pilot.cli:main"
screen-pilot-server = "screen_pilot.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/screen_pilot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
